+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Changed
 Added
    - Arity of custom sub-commands is optional and defaults to `{min=0,
      max=nil}`
+   - Module completion in `:Debugpy module` subcommand
 
 0.6.0 - 2022-08-07
 ##################